pyproject.toml
[project]
name = "konrad-mvp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9, <3.15"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.12,<3.0.0)",
    "pypdf (>=6.1.0,<7.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn[standard] (>=0.36.0,<0.37.0)",
    "langchain-community (>=0.3.29,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false